% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataframe.R
\name{createDataframe}
\alias{createDataframe}
\title{Create a DataFrame from PhotosynQ Project Data}
\usage{
createDataframe(project_info = "", project_data = "")
}
\arguments{
\item{project_info}{Object returned by getProjectInfo()}

\item{project_data}{Object returned by getProjectData()}
}
\description{
This function allows you to create a dataframe based on the
Project information and Project data receive data from PhotosynQ.
}
\examples{
createDataframe(project_info,project_data)
}
\keyword{Data}
\keyword{Project}
\keyword{dataframe}
