% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProject.R
\name{getProject}
\alias{getProject}
\title{Get Project data in a Data Frame from PhotosynQ}
\usage{
getProject(projectID = "", processedData = TRUE, rawTraces = FALSE)
}
\arguments{
\item{projectID}{The ID of your Project (Just copy the Project ID from the project page or your user page)}

\item{processedData}{(optional) Receive the processed data when set to TRUE, receive raw Data when set to FALSE. (raw data will increases data size significantly!)}

\item{rawTraces}{(optional) Adds raw traces to processed data. It is ignored when processedData is set to TRUE. (increases data size significantly!)}
}
\description{
Get Project data using the Project ID and create data frame(s).
}
\details{
This function produces a data frame for a PhotosynQ Project using the Project's ID. The ID can be found on the
Projects page on the PhotosynQ website. This function call includes the Project data as well as information.
In case multiple protocols were used, each protocol is in a separate data frame. By default the 
processed data is not included. The parameter rawTraces can be set to TRUE to include the rawTraces in combination with
the processed data. By default the traces are not received.
When the original data structure is needed, separate calles have to be used including getProjectInfo and getProjectData.
Note: Including the raw data will increase the data size significantly.
}
\examples{
getProject(1566)
}
\keyword{Data}
\keyword{Project}
