% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to PhotosynQ}
\usage{
login(email = "")
}
\arguments{
\item{email}{Your email address you use to login}
}
\description{
This function allows you to login to PhotosynQ and receive your authentication token.
The functions getProjectInfo(), getProjectData() and logout() require the token to
receive data from PhotosynQ
}
\examples{
login("john.doe@domain.com")
}
\keyword{login}
