% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logout}
\alias{logout}
\title{Logout from PhotosynQ}
\usage{
logout()
}
\description{
This function allows you to logout from PhotosynQ after you finished with your session.
}
\examples{
logout()
}
\keyword{logout}
