% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataframe.R
\name{createDataframe}
\alias{createDataframe}
\title{Generate a Data Frame from PhotosynQ project data}
\usage{
createDataframe(project_info = "", project_data = "")
}
\arguments{
\item{project_info}{Object returned by getProjectInfo()}

\item{project_data}{Object returned by getProjectData()}
}
\description{
This function allows you to create a  data frame based on the
Project information and project data receive from PhotosynQ.
}
\examples{
createDataframe(project_info,project_data)
}
\keyword{Project}
\keyword{data}
\keyword{frame}
