% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProject.R
\name{getProject}
\alias{getProject}
\title{Get Project data in a Data Frame from PhotosynQ}
\usage{
getProject(projectID = "", processedData = TRUE, rawTraces = FALSE)
}
\arguments{
\item{projectID}{The ID of your Project (Just copy the Project ID from the project page or your user page)}

\item{processedData}{(optional) Receive the processed data when set to TRUE, receive raw Data when set to FALSE. (raw data will increases data size significantly!)}

\item{rawTraces}{(optional) Adds raw traces to processed data. It is ignored when processedData is set to TRUE. (increases data size significantly!)}
}
\description{
This function allows you to receive data from PhotosynQ and convert it into a  data frame.
}
\examples{
getProject(1566)
}
\keyword{Data}
\keyword{Project}
