% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProject.R
\name{getProject}
\alias{getProject}
\title{Get Project Data in a DataFrame from PhotosynQ}
\usage{
getProject(email = "", projectURL = "")
}
\arguments{
\item{email}{Your email address you use to sign in}

\item{projectURL}{The URL of your Project (Just copy the Project page URL from your browser)}
}
\description{
This function allows you to receive data from PhotosynQ.
}
\examples{
getProject("john.doe@gmail.com","SecurePassword","https://photosynq.org/projects/getting-started-with-multispeq")
}
\keyword{Data}
\keyword{Project}
