% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectData.R
\name{getProjectData}
\alias{getProjectData}
\title{Get Project Data from PhotosynQ}
\usage{
getProjectData(projectID = "", processedData = TRUE,
  rawTraces = FALSE)
}
\arguments{
\item{projectID}{The ID of your Project (Just copy the Project ID from the project page or your user page)}

\item{processedData}{(optional) Receive the processed data when set to TRUE, receive raw Data when set to FALSE. (raw data will increases data size significantly!)}

\item{rawTraces}{(optional) Adds raw traces to processed data. It is ignored when processedData is set to TRUE. (increases data size significantly!)}
}
\description{
This function allows you to receive data from PhotosynQ for a specific project.
The revceived data is in the original JSON structure.
}
\examples{
getProjectData(1566)
}
\keyword{Data}
\keyword{Project}
