% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectData.R
\name{getProjectData}
\alias{getProjectData}
\title{Get Project Data from PhotosynQ}
\usage{
getProjectData(email = "", token = "", projectURL = "")
}
\arguments{
\item{email}{Your email address you use to sign in}

\item{token}{Your password you use to sign in}

\item{projectURL}{The URL of your Project (Just copy the Project page URL from your browser)}
}
\description{
This function allows you to receive data from PhotosynQ.
}
\examples{
getProjectData("john.doe@gmail.com","A67DHsajjshda78","https://photosynq.org/projects/getting-started-with-multispeq")
}
\keyword{Data}
\keyword{Project}
