% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectInfo.R
\name{getProjectInfo}
\alias{getProjectInfo}
\title{Get Project Information from PhotosynQ}
\usage{
getProjectInfo(projectID = "")
}
\arguments{
\item{projectID}{The ID of your Project (Just copy the Project ID from the project page or your user page)}
}
\description{
This function allows you to receive the information about a project from PhotosynQ.
}
\examples{
getProjectInfo(1566)
}
\keyword{Information}
\keyword{Project}
