% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectInfo.R
\name{getProjectInfo}
\alias{getProjectInfo}
\title{Get Project Information from PhotosynQ}
\usage{
getProjectInfo(email = "", token = "", projectURL = "")
}
\arguments{
\item{email}{Your email address you use to login}

\item{token}{Your password you use to sign in}

\item{projectURL}{The URL of your Project (Just copy the Project page URL from your browser)}
}
\description{
This function allows you to sign in to PhotosynQ and receive your authentication token.
}
\examples{
getProjectInfo("john.doe@gmail.com","A67DHsajjshda78","https://photosynq.org/projects/getting-started-with-multispeq")
}
\keyword{Information}
\keyword{Project}
