% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to PhotosynQ}
\usage{
login(email = "", url = photosynq.env$DEFAULT_API_DOMAIN)
}
\arguments{
\item{email}{Your email address you use to login}

\item{url}{(optional) Change the default URL to point to another instance}
}
\description{
Login to PhotosynQ to allow data access.
}
\details{
This function allows a user to login to PhotosynQ and start a session.
The functions getProjectInfo(), getProjectData(), getProject() and logout()
require a session started by login in.
Login is only required once at the beginning of a session.
}
\examples{
login("john.doe@domain.com")
}
\keyword{login}
