% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logout}
\alias{logout}
\title{Logout from 'PhotosynQ'}
\usage{
logout()
}
\value{
Session data is removed from the global variables. Nothing is
  returned
}
\description{
Logout from 'PhotosynQ' and end session
}
\details{
This function ends the current session and logs out the current user out from
'PhotosynQ'. Use the login function to sign in again and start a new session
if needed.
}
\examples{
logout()
}
\keyword{logout}
