% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logout}
\alias{logout}
\title{Logout from PhotosynQ}
\usage{
logout(token = "")
}
\arguments{
\item{token}{Your token from the initial login.}
}
\description{
This function allows you to logout from PhotosynQ after you are done with your session.
}
\examples{
logout("A67DHsajjshda78")
}
\keyword{logout}
